refactor: house external_fetch in shared/ package with core/shared module arrays
- Move common/external_fetch/ -> shared/external_fetch/; extract its DI
  bindings out of AppModule into a dedicated ExternalFetchModule.
- quickapp.core exposes core_module (App + Agent modules); quickapp.shared
  exposes shared_module (ExternalFetchModule). app_factory and the
  integration-test injector splice both arrays instead of registering
  each module individually.
- Align the external-URL design doc with the implemented attachment
  emission (reference_url + url on the url_attachments-supported branch).
