+++ src/quickapp/shared/external_fetch/external_fetch_settings.py
