+++ src/quickapp/shared/external_fetch/host_pattern_match.py
