+++ src/tests/unit_tests/file_transfer_tests/test_file_loader_service.py
@@ -6,15 +6,15 @@ from aidial_client.types.metadata import FileMetadata
 
 from quickapp.common.dial_settings import DialSettings
 from quickapp.common.exceptions import InvalidToolCallParameterException
-from quickapp.common.external_fetch.external_url_fetcher import (
+from quickapp.common.state_holder import StateHolder
+from quickapp.dial_core_services.dial_downloader import DialDownloader
+from quickapp.file_transfer._file_loader_service import FileLoaderService
+from quickapp.shared.external_fetch.external_url_fetcher import (
     ExternalFetchDisabledError,
     ExternalFetchError,
     ExternalUrlFetcher,
     FetchedBytes,
 )
-from quickapp.common.state_holder import StateHolder
-from quickapp.dial_core_services.dial_downloader import DialDownloader
-from quickapp.file_transfer._file_loader_service import FileLoaderService
 
 DIAL_URL = "https://dial.example.com"
