+++ src/tests/unit_tests/shared/test_host_pattern_match.py
@@ -1,6 +1,6 @@
 import pytest
 
-from quickapp.common.external_fetch.host_pattern_match import match_host
+from quickapp.shared.external_fetch.host_pattern_match import match_host
 
 
 @pytest.mark.parametrize(
